树展示

+++ familytree/src/main/java/com/starfire/familytree/vo/OrgChartItemVO.java
@@ -2,10 +2,8 @@ package com.starfire.familytree.vo;
 
 import lombok.Data;
 
-import java.util.List;
-
 /**
- * 组织图表项目
+ * 组织图表项目,封装basicprimitives item对象，id和parent 类型为long 无法正常表示关系
  * @version 1.0
  * @author: Luzh
  * @date: 2019/8/30 8:45
@@ -13,8 +11,9 @@ import java.util.List;
 @Data
 public class OrgChartItemVO {
 
-    private Long  id;
-    private Long  parent;
+
+    private Integer  id;
+    private Integer  parent;
     private  String title;
     private  String description;
     private  String image;
